+++ product-service/src/main/java/com/structura/steel/productservice/controller/ProductController.java
@@ -9,6 +9,8 @@ import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -62,8 +64,8 @@ public class ProductController {
     }
 
     @GetMapping("/{id}/weight")
-    public ResponseEntity<Double> getProductWeight(@PathVariable Long id) {
-        double weight = productService.calculateWeight(id);
+    public ResponseEntity<BigDecimal> getProductWeight(@PathVariable Long id) {
+        BigDecimal weight = productService.calculateWeight(id);
         return ResponseEntity.ok(weight);
     }
 }
